working state

+++ modular_zubbers/code/datums/components/dollification.dm
@@ -1,14 +1,3 @@
-/// Item handling
-
-/obj/item/clothing/mask/muzzle/ring/doll
-	name = "fused opened lips"
-	desc = "Your lips are stuck open like this. No matter how much you prod at them, you can't close them!"
-
-/obj/item/clothing/shoes/fancy_heels/doll
-	name = "attached high heels"
-	desc = "These obsidian heels refuse to detatch from your feet. They feel like a part of you now."
-	greyscale_colors = "#1b1b1b"
-
 /// Admin Smite
 /datum/smite/dollify
 	name = "Dollify"
@@ -38,7 +27,7 @@
 
 	/// Item handling
 	var/obj/item/clothing/mask/muzzle/ring/doll/gag
-	var/obj/item/clothing/shoes/fancy_heels/doll/heels
+	var/obj/item/clothing/shoes/fancy_heels/black/doll/heels
 
 	/// Cached random starting point. Purely for flavor.
 	var/origin
@@ -57,6 +46,7 @@
 
 /datum/component/dollification/Initialize()
 	. = ..()
+	var/list/orgins = list("hands", "crotch", "cheek", "butt", "thigh", "chest")
 	doll = parent
 	if(!iscarbon(doll))
 		return COMPONENT_INCOMPATIBLE
@@ -70,10 +60,10 @@
 	new_name = generate_name()
 
 	RegisterSignal(doll, COMSIG_ATOM_EXAMINE, PROC_REF(on_doll_examine))
-	var/list/orgins = list("hands", "crotch", "cheek", "butt", "thigh", "chest")
 	origin = pick(orgins)
 	to_chat(doll, span_purple("<b>You feel something slick smudge across your [origin]; a glossy smear visibly spreading in every direction!</b>"))
 	progress_stage()
+	uncurable = permanency_prompt()
 	START_PROCESSING(SSobj, src)
 
 /datum/component/dollification/Destroy()
@@ -89,17 +79,16 @@
 	var/probability = 5
 	if(doll.stat >= HARD_CRIT) // You shouldn't progress if you're fucked up.
 		return
-	if(stage == 0 || uncurable) // Kills processing when it's not needed anymore.
-		STOP_PROCESSING(SSobj, src)
-		return
 
-	if(doll.reagents.has_reagent(/datum/reagent/dinitrogen_plasmide, 20)) // The cure
+	if(doll.reagents.has_reagent(/datum/reagent/dinitrogen_plasmide, 20) && !uncurable) // The cure
 		qdel(src)
 
 	if(!COOLDOWN_FINISHED(src, doll_grace_period)) // Have a little bit of time between stages.
 		return
+
 	if(dollified)
 		return
+
 	/// Progresses the infestation based on a probability and runs through each stage's transformation descriptions before moving on.
 	if(SPT_PROB(probability, seconds_per_tick))
 		var/picked_text = pick_n_take(infection_text)
@@ -113,11 +102,11 @@
 	var/number = pick(rand(1000,9999))
 
 	if(isnull(doll.job))
-		job = "Doll"
+		job = "Standard"
 	else
 		job = doll.job
 
-	return "[job] #[number]"
+	return "[job] Doll #[number]"
 
 /// Examine handler
 /datum/component/dollification/proc/on_doll_examine(atom/source, mob/user, list/examine_list)
@@ -127,13 +116,52 @@
 		return
 	examine_list += span_purple(examine_text)
 
-/// Infection stage handling. Forced is an extra logic check so that undroppable gags are not duplicating.
-/datum/component/dollification/proc/progress_stage(level, forced = FALSE)
-	if(!isnull(level))
-		stage = level
-	else
-		stage++
+/// Infection stage handling for item spawning.
+/datum/component/dollification/proc/progress_stage()
+	stage++
 
+	switch(stage)
+		if(2)
+			if(doll.get_item_by_slot(ITEM_SLOT_FEET))
+				doll.dropItemToGround(doll.shoes, TRUE)
+			heels = new(doll.loc)
+			ADD_TRAIT(heels, TRAIT_NODROP, TRAIT_DOLLIFICATION)
+			doll.equip_to_slot_if_possible(heels, ITEM_SLOT_FEET, TRUE, TRUE, TRUE)
+
+		if(3)
+			if(doll.get_item_by_slot(ITEM_SLOT_MASK))
+				doll.dropItemToGround(doll.wear_mask, TRUE)
+
+			gag = new(doll.loc)
+			ADD_TRAIT(gag, TRAIT_NODROP, TRAIT_DOLLIFICATION)
+			doll.equip_to_slot_if_possible(gag, ITEM_SLOT_MASK, TRUE, TRUE, TRUE)
+		if(4)
+			finalize_doll()
+
+	generate_stage_flavour()
+	COOLDOWN_START(src, doll_grace_period, 10 SECONDS)
+
+/// Ask them if they want this to be permanent.
+/datum/component/dollification/proc/permanency_prompt()
+	var/action
+	action = tgui_alert(doll, "Warning: This is unrevertable in character unless you use OOC Safe Word", "Permanent Dollification?", list("No", "Yes"))
+	if(action == "No")
+		return FALSE
+	if(action == "Yes")
+		return TRUE
+
+/// Handle the final stages and change your name.
+/datum/component/dollification/proc/finalize_doll()
+	dollified = TRUE
+	doll.fully_replace_character_name(old_name, new_name)
+	if(uncurable)
+		to_chat(doll, span_purple("<b>You can feel your entire body succumbing to its fate. The glossy sheen not just soaking through your flesh, but muscle and bone. <b>Fuck</b>, You're stuck like this... and it feels permanent.</b>"))
+		STOP_PROCESSING(SSobj, src) // Not needed anymore!
+	else
+		to_chat(doll, span_purple("<b>The changes have finished and you're now settling into your new roll. Perhaps there's a way to escape?</b>"))
+
+// Handles all descriptions and long transformation text in its own seperate proc.
+/datum/component/dollification/proc/generate_stage_flavour()
 	switch(stage)
 		if(1)
 			examine_text = "There is a festing splotch of smoothness slowly encompassing their [origin]"
@@ -150,52 +178,14 @@
 							"You're starting to walk in a seductive saunter. You just can't seem to keep your hips straight!",
 							"Your heeled feet clack on the hard station floor. Each click a little shockwave tingling up your rubberized legs.",
 							"The creeping elastic yearns to assimilate your flesh. You're becoming more <b>thing</b> than person as the long seconds tick by.")
-			if(forced)
-				return
-
-			if(doll.get_item_by_slot(ITEM_SLOT_FEET) && !forced)
-				doll.dropItemToGround(doll.shoes, TRUE)
-
-			heels = new(doll.loc)
-			ADD_TRAIT(heels, TRAIT_NODROP, TRAIT_DOLLIFICATION)
-			doll.equip_to_slot_if_possible(heels, ITEM_SLOT_FEET, TRUE, TRUE, TRUE)
-
 		if(3)
 			to_chat(doll, span_purple("<b>Oooawwhh? Tension floods through your lips, forcing them into a signature O, The elastic has finally crept over the rest of your head.</b>"))
 			examine_text = "A rubbery sheen coats their entire body."
 			infection_text = list(
-							"You try to flex your jaw and close your mouth. It only gapes wider.",
+							"You try to flex your jaw and close your mouth. It only gapes wider...",
 							"Lubrication fills your open mouth and begins to leak out.",
 							"Your mouth is tingling as the rubbery materal begins spreading inwards.",
 							"You're feeling more and more rewired- trained- alert to any sudden demands.",
 							"Your name is slipping from you like sand. <b>Shit</b>, what-t was it again? [new_name]? No- it's [old_name], right?")
-			if(forced)
-				return
-
-			if(doll.get_item_by_slot(ITEM_SLOT_MASK))
-				doll.dropItemToGround(doll.wear_mask, TRUE)
-
-			gag = new(doll.loc)
-			ADD_TRAIT(gag, TRAIT_NODROP, TRAIT_DOLLIFICATION)
-			doll.equip_to_slot_if_possible(gag, ITEM_SLOT_MASK, TRUE, TRUE, TRUE)
-
 		if(4)
 			examine_text = "Every movement is a subtle squrk: a walking, talking, breathing doll!"
-			permanency_prompt()
-
-	COOLDOWN_START(src, doll_grace_period, 10 SECONDS)
-
-/// Ask them if they want this to be permanent. Also changes their name when they reach this point.
-/datum/component/dollification/proc/permanency_prompt()
-	var/action
-	doll.fully_replace_character_name(old_name, new_name)
-	dollified = TRUE
-	action = tgui_alert(doll, "Warning: This is unrevertable through ordinary means", "Do you wish to become a doll permanently?", list("No", "Yes"))
-	if(action == "No")
-		progress_stage(3, TRUE)
-		to_chat(doll, span_purple("The changes have finished and you're now settling into your new roll. Perhaps there's a way to escape?"))
-
-	if(action == "Yes")
-		uncurable = TRUE
-		to_chat(doll, span_purple("You can feel your entire body succumbing to its fate. The glossy sheen not just soaking through your flesh, but muscle and bone. <b>Fuck</b>, You're stuck like this... and it feels permanent."))
-
